requirements/base.txt
# Django
Django==5.1.6
django-environ==0.12.0
django-redis==5.4.0
django-crispy-forms==2.3
django-filter==25.1
django-import-export==4.3.7

# Others
celery==5.4.0
psycopg2-binary==2.9.10
crispy-bootstrap5==2024.10
cryptography==44.0.1
openpyxl==3.1.5